backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3

# Core dependencies (from root requirements.txt)
streamlit==1.40.1
groq==0.11.0
azure-cognitiveservices-speech==1.40.0
redis==5.2.0
sounddevice==0.5.1
numpy==2.1.3
scipy==1.14.1
